Change S3 bucket removal policy to retain
The removal policy for the S3 bucket in the AWS stack configuration file has been updated. Originally, the policy was set to 'destroy', meaning all contents of the bucket were permanently deleted upon removal. Now, with the policy changed to 'retain', data will be kept even if the bucket is deleted.